move the integration functions into their own file
